edited the files to incorporate new code.
