+++ server/app.py
@@ -1,3 +1,21 @@
-from.routes.auth import auth_blueprint
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_jwt_extended import JWTManager
+from config import Config
+from models import db
+from auth import auth_blueprint
+
+app = Flask(__name__)
+app.config.from_object(Config)
+db.init_app(app)
+jwt = JWTManager(app)
 
 app.register_blueprint(auth_blueprint, url_prefix='/api/auth')
+
+if __name__ == "__main__":
+    db.create_all()
+    db.session.add(User(full_name="Bruce Wayne", username="batman"))
+    db.session.add(User(full_name="Ann Takamaki", username="panther"))
+    db.session.add(User(full_name="Jester Lavore", username="little_sapphire"))
+    db.session.commit()
+    app.run()
