+++ server/models/user.py
@@ -5,7 +5,8 @@ db = SQLAlchemy()
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(80), unique=True, nullable=False)
-    password = db.Column(db.String(120), nullable=False)
+    full_name = db.Column(db.String(120), nullable=False)
 
-    def __repr__(self):
-        return f"User('{self.username}')"
+    def check_password(self, password):
+        from hmac import compare_digest
+        return compare_digest(password, "password")
